refactor(기능): 클래스 이름 변경(DishFinder -> OrderMaker), 코드 단순화
